feat: fixed datepicker of scheduleAddView
